requirements.txt
python-telegram-bot==21.6
telethon==1.36.0
python-dotenv==1.0.1
cryptg==0.4.0
